fixed file size bug on 32-bit platforms

+++ bigBuffer.h
@@ -22,6 +22,7 @@
 #define BIG_BUFFER_H
 
 #include <zip.h>
+#include <unistd.h>
 
 #include <vector>
 
@@ -33,18 +34,18 @@ private:
 
     chunks_t chunks;
 
-    static off_t zipUserFunctionCallback(void *state, void *data, size_t len, enum zip_source_cmd cmd);
+    static ssize_t zipUserFunctionCallback(void *state, void *data, size_t len, enum zip_source_cmd cmd);
 public:
-    ssize_t len;
+    off64_t len;
 
     BigBuffer();
     BigBuffer(struct zip *z, int nodeId, ssize_t length);
     ~BigBuffer();
 
-    int read(char *buf, size_t size, off_t offset) const;
-    int write(const char *buf, size_t size, off_t offset);
+    int read(char *buf, size_t size, off64_t offset) const;
+    int write(const char *buf, size_t size, off64_t offset);
     int saveToZip(struct zip *z, const char *fname, bool newFile, int index);
-    int truncate(off_t offset);
+    int truncate(off64_t offset);
 };
 
 #endif
